Move the VulkanRenderManager stuff to thin3d, fits in better there.
