+++ Common/Common.vcxproj
@@ -249,8 +249,6 @@
     <ClInclude Include="Vulkan\VulkanImage.h" />
     <ClInclude Include="Vulkan\VulkanLoader.h" />
     <ClInclude Include="Vulkan\VulkanMemory.h" />
-    <ClInclude Include="Vulkan\VulkanQueueRunner.h" />
-    <ClInclude Include="Vulkan\VulkanRenderManager.h" />
     <ClInclude Include="x64Analyzer.h" />
     <ClInclude Include="x64Emitter.h" />
   </ItemGroup>
@@ -322,8 +320,6 @@
     <ClCompile Include="Vulkan\VulkanImage.cpp" />
     <ClCompile Include="Vulkan\VulkanLoader.cpp" />
     <ClCompile Include="Vulkan\VulkanMemory.cpp" />
-    <ClCompile Include="Vulkan\VulkanQueueRunner.cpp" />
-    <ClCompile Include="Vulkan\VulkanRenderManager.cpp" />
     <ClCompile Include="x64Analyzer.cpp" />
     <ClCompile Include="x64Emitter.cpp" />
   </ItemGroup>
