+++ ext/native/native.vcxproj.filters
@@ -320,6 +320,12 @@
     <ClInclude Include="gfx_es2\draw_text_android.h">
       <Filter>gfx</Filter>
     </ClInclude>
+    <ClInclude Include="thin3d\VulkanQueueRunner.h">
+      <Filter>thin3d</Filter>
+    </ClInclude>
+    <ClInclude Include="thin3d\VulkanRenderManager.h">
+      <Filter>thin3d</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="gfx\gl_debug_log.cpp">
@@ -780,6 +786,12 @@
     </ClCompile>
     <ClCompile Include="math\dataconv.cpp">
       <Filter>math</Filter>
+    </ClCompile>
+    <ClCompile Include="thin3d\VulkanQueueRunner.cpp">
+      <Filter>thin3d</Filter>
+    </ClCompile>
+    <ClCompile Include="thin3d\VulkanRenderManager.cpp">
+      <Filter>thin3d</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
